🐛 fix int-map key types (#3829)
We currently don't have deep support for int-maps, since we don't have
any use-cases for them yet. That said, they have been supported in the
underlying stack for when we need them. Unfortunately a few of the
supporting elements have been out of sync.

This change aligns data conversions in llx to use int64 maps. Ditto for
`isTruthy` in rawdata. There are a few more to tackle...

+++ llx/data_conversions.go
@@ -286,7 +286,7 @@ func stringmap2result(value interface{}, typ types.Type) (*Primitive, error) {
 }
 
 func intmap2result(value interface{}, typ types.Type) (*Primitive, error) {
-	m, ok := value.(map[int32]interface{})
+	m, ok := value.(map[int64]interface{})
 	if !ok {
 		return nil, errInvalidConversion(value, typ)
 	}
@@ -294,7 +294,7 @@ func intmap2result(value interface{}, typ types.Type) (*Primitive, error) {
 	ct := typ.Child()
 	var err error
 	for k, v := range m {
-		res[strconv.FormatInt(int64(k), 10)], err = raw2primitive(v, ct)
+		res[strconv.FormatInt(k, 10)], err = raw2primitive(v, ct)
 		if err != nil {
 			return nil, err
 		}
